Serverside websocket, clientside websocket
Added functionality for websockets, on the serverside, and also on the
clientside. Now we can keep track of active users on the serverside (
game.esj ), inactive users get disconnected from the server after 10
seconds. On the clientside ( if you connect with your phone ) you can
send commands to the server (left / right) for the game dieCurve. Also
added somegame functionality already